# frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Resume Builder API specification",
    "description": "API for Resume Builder",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8080/api/v1", "description": "Local server" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/sections": {
      "post": {
        "tags": ["Resume Sections"],
        "operationId": "createSection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SectionRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/sections/{id}": {
      "get": {
        "tags": ["Resume Sections"],
        "operationId": "getSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SectionResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Resume Sections"],
        "operationId": "updateSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SectionRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/sections/{id}/delete": {
      "post": {
        "tags": ["Resume Sections"],
        "operationId": "deleteSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/section-items": {
      "post": {
        "tags": ["section-item-controller"],
        "operationId": "createSectionItem",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SectionItemRequest" }
            },
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/section-items/{id}": {
      "get": {
        "tags": ["section-item-controller"],
        "operationId": "getSectionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SectionItemResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["section-item-controller"],
        "operationId": "updateSectionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SectionItemRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/section-items/{id}/delete": {
      "post": {
        "tags": ["section-item-controller"],
        "operationId": "deleteSectionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/resumes": {
      "get": {
        "tags": ["Resume"],
        "operationId": "getAllResumes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "lastModifiedDate" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseResumeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Resume"],
        "operationId": "createResume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResumeRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/resumes/{id}": {
      "get": {
        "tags": ["Resume"],
        "operationId": "getResume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeDetailResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Resume"],
        "operationId": "updateResume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResumeRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/resumes/{id}/uploadPicture": {
      "post": {
        "tags": ["Resume"],
        "operationId": "uploadResumePicture",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/resumes/{id}/delete": {
      "post": {
        "tags": ["Resume"],
        "operationId": "deleteResume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SectionItemRequest": {
        "required": ["data", "latexMethodId", "type"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "sectionId": { "type": "integer", "format": "int64" },
          "latexMethodId": { "type": "integer", "format": "int64" },
          "type": { "type": "string" },
          "itemOrder": { "type": "integer", "format": "int32" },
          "alignment": {
            "type": "string",
            "enum": ["LEFT", "CENTER", "RIGHT", "JUSTIFIED"]
          },
          "data": {
            "type": "object",
            "additionalProperties": { "type": "object" }
          }
        }
      },
      "SectionRequest": {
        "required": ["showTitle", "title"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "resumeId": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "showTitle": { "type": "boolean" },
          "sectionItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionItemRequest" }
          }
        }
      },
      "ResumeRequest": {
        "required": ["title"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "sections": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionRequest" }
          }
        }
      },
...
      "LatexMethodResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "type": {
            "type": "string",
            "enum": [
              "TITLE",
              "WORK_EXPERIENCE",
              "TEXTBOX",
              "SKILL",
              "EDUCATION",
              "SKILL_BOXES",
              "PICTURE",
              "CONTACT"
            ]
          },
          "name": { "type": "string" },
          "method": { "type": "string" }
        }
      },
      "SectionItemResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "latexMethod": { "$ref": "#/components/schemas/LatexMethodResponse" },
          "type": { "type": "string" },
          "itemOrder": { "type": "integer", "format": "int32" },
          "alignment": {
            "type": "string",
            "enum": ["LEFT", "CENTER", "RIGHT", "JUSTIFIED"]
          },
          "data": {
            "type": "object",
            "additionalProperties": { "type": "object" }
          }
        }
      },
      "SectionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "showTitle": { "type": "boolean" },
          "sectionItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionItemResponse" }
          }
        }
      },
      "PageResponseResumeResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ResumeResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      },
      "ResumeResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "picture": { "type": "string", "format": "byte" },
          "createdDate": { "type": "string" },
          "lastModifiedDate": { "type": "string" }
        }
      },
      "ResumeDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "picture": { "type": "string", "format": "byte" },
          "createdDate": { "type": "string" },
          "lastModifiedDate": { "type": "string" },
          "sections": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionResponse" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}